Added attribute inputValue
Added attribute inputValue to fill input value without triggering a search. Example: Editing a reactiveform with multiple addresses field, each address formgroup get inserted and populated with data returned from google maps search, and in the input value i only show a text with the resume of the address -> formatted_address, when i open the form for editing i only want to show the formatted_address in the search box without triggering a search, if the value is changed a new search is triggered.